Extract shared code for workflow components into a common base class
