+++ packages/web-client/app/components/card-pay/deposit-workflow/index.ts
@@ -1,5 +1,3 @@
-import Component from '@glimmer/component';
-import { getOwner } from '@ember/application';
 import {
   Workflow,
   Milestone,
@@ -17,9 +15,8 @@ import { action } from '@ember/object';
 import { currentNetworkDisplayInfo as c } from '@cardstack/web-client/utils/web3-strategies/network-display-info';
 import { capitalize } from '@ember/string';
 import RouterService from '@ember/routing/router-service';
-import WorkflowPersistence from '@cardstack/web-client/services/workflow-persistence';
-import { tracked, cached } from '@glimmer/tracking';
 import { standardCancelationPostables } from '@cardstack/web-client/models/workflow/cancelation-helpers';
+import RestorableWorkflowComponent from '../restorable-workflow-component';
 
 const FAILURE_REASONS = {
   DISCONNECTED: 'DISCONNECTED',
@@ -256,30 +253,12 @@ class DepositWorkflow extends Workflow {
   }
 }
 
-class DepositWorkflowComponent extends Component {
+class DepositWorkflowComponent extends RestorableWorkflowComponent<DepositWorkflow> {
   @service declare layer1Network: Layer1Network;
   @service declare layer2Network: Layer2Network;
-  @service declare workflowPersistence: WorkflowPersistence;
-  @service declare router: RouterService;
 
-  @tracked isInitializing = true;
-
-  get workflowPersistenceId() {
-    return this.router.currentRoute.queryParams['flow-id']!;
-  }
-
-  @cached
-  get workflow() {
-    return new DepositWorkflow(getOwner(this), this.workflowPersistenceId);
-  }
-  constructor(owner: unknown, args: {}) {
-    super(owner, args);
-    this.restore();
-  }
-
-  async restore() {
-    await this.workflow.restore();
-    this.isInitializing = false;
+  get workflowClass() {
+    return DepositWorkflow;
   }
 
   @action onDisconnect() {
