+++ packages/web-client/app/components/card-pay/issue-prepaid-card-workflow/index.ts
@@ -1,9 +1,6 @@
-import Component from '@glimmer/component';
-import { getOwner } from '@ember/application';
 import { inject as service } from '@ember/service';
 
 import Layer2Network from '@cardstack/web-client/services/layer2-network';
-import WorkflowPersistence from '@cardstack/web-client/services/workflow-persistence';
 import { action } from '@ember/object';
 import RouterService from '@ember/routing/router-service';
 
@@ -23,13 +20,13 @@ import {
   conditionalCancelationMessage,
 } from '@cardstack/web-client/models/workflow';
 
-import { cached, tracked } from '@glimmer/tracking';
 import {
   convertAmountToNativeDisplay,
   fromWei,
   spendToUsd,
 } from '@cardstack/cardpay-sdk';
 import { standardCancelationPostables } from '@cardstack/web-client/models/workflow/cancelation-helpers';
+import RestorableWorkflowComponent from '../restorable-workflow-component';
 
 export const faceValueOptions = [500, 1000, 2500, 5000, 10000, 50000];
 
@@ -362,33 +359,11 @@ class IssuePrepaidCardWorkflow extends Workflow {
   }
 }
 
-class IssuePrepaidCardWorkflowComponent extends Component {
+class IssuePrepaidCardWorkflowComponent extends RestorableWorkflowComponent<IssuePrepaidCardWorkflow> {
   @service declare layer2Network: Layer2Network;
-  @service declare workflowPersistence: WorkflowPersistence;
-  @service declare router: RouterService;
 
-  @tracked isInitializing = true;
-
-  get workflowPersistenceId() {
-    return this.router.currentRoute.queryParams['flow-id']!;
-  }
-
-  @cached
-  get workflow() {
-    return new IssuePrepaidCardWorkflow(
-      getOwner(this),
-      this.workflowPersistenceId
-    );
-  }
-
-  constructor(owner: unknown, args: {}) {
-    super(owner, args);
-    this.restore();
-  }
-
-  async restore() {
-    await this.workflow.restore();
-    this.isInitializing = false;
+  get workflowClass() {
+    return IssuePrepaidCardWorkflow;
   }
 
   @action onDisconnect() {
