+++ packages/web-client/app/components/card-space/create-space-workflow/index.ts
@@ -1,11 +1,8 @@
-import Component from '@glimmer/component';
-import { getOwner } from '@ember/application';
 import { inject as service } from '@ember/service';
 import { action } from '@ember/object';
-import { tracked, cached } from '@glimmer/tracking';
+import { tracked } from '@glimmer/tracking';
 import RouterService from '@ember/routing/router-service';
 import Layer2Network from '@cardstack/web-client/services/layer2-network';
-import WorkflowPersistence from '@cardstack/web-client/services/workflow-persistence';
 import HubAuthentication from '@cardstack/web-client/services/hub-authentication';
 import { currentNetworkDisplayInfo as c } from '@cardstack/web-client/utils/web3-strategies/network-display-info';
 import {
@@ -20,6 +17,7 @@ import {
   conditionalCancelationMessage,
 } from '@cardstack/web-client/models/workflow';
 import { standardCancelationPostables } from '@cardstack/web-client/models/workflow/cancelation-helpers';
+import RestorableWorkflowComponent from '../../card-pay/restorable-workflow-component';
 
 const FAILURE_REASONS = {
   L2_DISCONNECTED: 'L2_DISCONNECTED',
@@ -243,31 +241,13 @@ class CreateSpaceWorkflow extends Workflow {
   }
 }
 
-class CreateSpaceWorkflowComponent extends Component {
+export default class CreateSpaceWorkflowComponent extends RestorableWorkflowComponent<CreateSpaceWorkflow> {
   @service declare layer2Network: Layer2Network;
-  @service declare workflowPersistence: WorkflowPersistence;
-  @service declare router: RouterService;
 
   @tracked detailsEditFormShown: boolean = true;
-  @tracked isInitializing = true;
 
-  get workflowPersistenceId() {
-    return this.router.currentRoute.queryParams['flow-id']!;
-  }
-
-  @cached
-  get workflow() {
-    return new CreateSpaceWorkflow(getOwner(this), this.workflowPersistenceId);
-  }
-
-  constructor(owner: unknown, args: {}) {
-    super(owner, args);
-    this.restore();
-  }
-
-  async restore() {
-    await this.workflow.restore();
-    this.isInitializing = false;
+  get workflowClass() {
+    return CreateSpaceWorkflow;
   }
 
   @action onDisconnect() {
@@ -278,5 +258,3 @@ class CreateSpaceWorkflowComponent extends Component {
     this.workflow?.cancel(FAILURE_REASONS.L2_ACCOUNT_CHANGED);
   }
 }
-
-export default CreateSpaceWorkflowComponent;
